+++ amazon-chime-sdk/src/test/java/com/amazonaws/services/chime/sdk/meetings/ingestion/DefaultEventReporterTests.kt
@@ -5,7 +5,6 @@
 
 package com.amazonaws.services.chime.sdk.meetings.ingestion
 
-import com.amazonaws.services.chime.sdk.meetings.analytics.EventAttributeName
 import com.amazonaws.services.chime.sdk.meetings.analytics.EventName
 import com.amazonaws.services.chime.sdk.meetings.internal.ingestion.SDKEvent
 import com.amazonaws.services.chime.sdk.meetings.utils.logger.Logger
@@ -22,6 +21,9 @@ class DefaultEventReporterTests {
     private lateinit var ingestionConfiguration: IngestionConfiguration
 
     @MockK
+    private lateinit var eventClientConfiguration: EventClientConfiguration
+
+    @MockK
     private lateinit var eventBuffer: EventBuffer
 
     @MockK
@@ -30,11 +32,23 @@ class DefaultEventReporterTests {
     @MockK
     private lateinit var logger: Logger
 
+    private val flushIntervalMs = 200L
+    private val metadataAttributeKey1 = "metadataAttributeKey1"
+    private val metadataAttributeValue1 = 1
+    private val metadataAttributeKey2 = "metadataAttributeKey2"
+    private val metadataAttributeValue2 = "text"
+    private val metadataAttributes = mapOf(
+        metadataAttributeKey1 to metadataAttributeValue1,
+        metadataAttributeKey2 to metadataAttributeValue2
+    )
+
     @Before
     fun setup() {
         MockKAnnotations.init(this, relaxed = true)
-        every { ingestionConfiguration.flushSize } returns 20
-        every { ingestionConfiguration.flushIntervalMs } returns 200
+        every { ingestionConfiguration.disabled } returns false
+        every { ingestionConfiguration.flushIntervalMs } returns flushIntervalMs
+        every { ingestionConfiguration.clientConfiguration } returns eventClientConfiguration
+        every { eventClientConfiguration.metadataAttributes } returns metadataAttributes
     }
 
     @Test
@@ -53,8 +67,6 @@ class DefaultEventReporterTests {
 
     @Test
     fun `constructor should invoke timer schedule if enabled`() {
-        every { ingestionConfiguration.disabled } returns false
-
         DefaultEventReporter(
             ingestionConfiguration,
             eventBuffer,
@@ -62,27 +74,22 @@ class DefaultEventReporterTests {
             timer
         )
 
-        verify(exactly = 1) { timer.scheduleAtFixedRate(any(), any<Long>(), any()) }
+        verify(exactly = 1) { timer.scheduleAtFixedRate(any(), flushIntervalMs, flushIntervalMs) }
     }
 
     @Test
-    fun `report should call eventBuffer add`() {
-        val meetingId = "meetingId"
-        val attendeeId = "attendeeId"
-        val joinToken = ""
-        every { ingestionConfiguration.disabled } returns false
-        every { ingestionConfiguration.clientConfiguration } returns MeetingEventClientConfiguration(joinToken, meetingId, attendeeId)
-
+    fun `report should add configuration common attributes`() {
         val defaultEventReporter = DefaultEventReporter(
             ingestionConfiguration,
             eventBuffer,
             logger
         )
+
         defaultEventReporter.report(SDKEvent(EventName.meetingFailed, mutableMapOf()))
 
         verify(exactly = 1) { eventBuffer.add(match {
-            it.eventAttributes[EventAttributeName.meetingId] == meetingId &&
-                    it.eventAttributes[EventAttributeName.attendeeId] == attendeeId }
+            it.eventAttributes[metadataAttributeKey1] == metadataAttributeValue1 &&
+                    it.eventAttributes[metadataAttributeKey2] == metadataAttributeValue2 }
         ) }
     }
 }